+++ wp-job-portal/modules/category/controller.php
@@ -14,7 +14,7 @@ class WPJOBPORTALCategoryController {
 
     function handleRequest() {
         $layout = WPJOBPORTALrequest::getLayout('wpjobportallt', null, 'categories');
-        if (self::canaddfile()) {
+        if (self::canaddfile($layout)) {
             switch ($layout) {
                 case 'admin_categories':
                     WPJOBPORTALincluder::getJSModel('category')->getAllCategories();
@@ -34,15 +34,19 @@ class WPJOBPORTALCategoryController {
         }
     }
 
-    function canaddfile() {
+    function canaddfile($layout) {
         $nonce_value = WPJOBPORTALrequest::getVar('wpjobportal_nonce');
         if ( wp_verify_nonce( $nonce_value, 'wpjobportal_nonce') ) {
             if (isset($_POST['form_request']) && $_POST['form_request'] == 'wpjobportal')
                 return false;
             elseif (isset($_GET['action']) && $_GET['action'] == 'wpjobportaltask')
                 return false;
-            else
+            else{
+                if(!is_admin() && strpos($layout, 'admin_') === 0){
+                    return false;
+                }
                 return true;
+            }
         }
     }
 
@@ -50,6 +54,9 @@ class WPJOBPORTALCategoryController {
         $nonce = WPJOBPORTALrequest::getVar('_wpnonce');
         if (! wp_verify_nonce( $nonce, 'wpjobportal_category_nonce') ) {
              die( 'Security check Failed' );
+        }
+        if (!current_user_can('manage_options')) { //only admin can DO it.
+            return false;
         }
         $data = WPJOBPORTALrequest::get('post');
         $result = WPJOBPORTALincluder::getJSModel('category')->storeCategory($data);
@@ -65,6 +72,9 @@ class WPJOBPORTALCategoryController {
         if (! wp_verify_nonce( $nonce, 'wpjobportal_category_nonce') ) {
              die( 'Security check Failed' );
         }
+        if (!current_user_can('manage_options')) { //only admin can DO it.
+            return false;
+        }
         $ids = WPJOBPORTALrequest::getVar('wpjobportal-cb');
         $result = WPJOBPORTALincluder::getJSModel('category')->deleteCategories($ids);
         $msg = WPJOBPORTALMessages::getMessage($result, 'category');
@@ -78,6 +88,9 @@ class WPJOBPORTALCategoryController {
         $nonce = WPJOBPORTALrequest::getVar('_wpnonce');
         if (! wp_verify_nonce( $nonce, 'wpjobportal_category_nonce') ) {
              die( 'Security check Failed' );
+        }
+        if (!current_user_can('manage_options')) { //only admin can DO it.
+            return false;
         }
         $pagenum = WPJOBPORTALrequest::getVar('pagenum');
         $ids = WPJOBPORTALrequest::getVar('wpjobportal-cb');
@@ -95,6 +108,9 @@ class WPJOBPORTALCategoryController {
         $nonce = WPJOBPORTALrequest::getVar('_wpnonce');
         if (! wp_verify_nonce( $nonce, 'wpjobportal_category_nonce') ) {
              die( 'Security check Failed' );
+        }
+        if (!current_user_can('manage_options')) { //only admin can DO it.
+            return false;
         }
         $pagenum = WPJOBPORTALrequest::getVar('pagenum');
         $ids = WPJOBPORTALrequest::getVar('wpjobportal-cb');
@@ -116,6 +132,9 @@ class WPJOBPORTALCategoryController {
         $nonce = WPJOBPORTALrequest::getVar('_wpnonce');
         if (! wp_verify_nonce( $nonce, 'wpjobportal_category_nonce') ) {
              die( 'Security check Failed' );
+        }
+        if (!current_user_can('manage_options')) { //only admin can DO it.
+            return false;
         }
         $post = WPJOBPORTALrequest::get('post');
         if($post['task'] == 'unpublish'){
