+++ wp-job-portal/modules/highesteducation/controller.php
@@ -16,7 +16,7 @@ class WPJOBPORTALHighesteducationController {
 
     function handleRequest() {
         $layout = WPJOBPORTALrequest::getLayout('wpjobportallt', null, 'highesteducations');
-        if (self::canaddfile()) {
+        if (self::canaddfile($layout)) {
             switch ($layout) {
                 case 'admin_highesteducations':
                     WPJOBPORTALincluder::getJSModel('highesteducation')->getAllHighestEducations();
@@ -36,15 +36,19 @@ class WPJOBPORTALHighesteducationController {
         }
     }
 
-    function canaddfile() {
+    function canaddfile($layout) {
         $nonce_value = WPJOBPORTALrequest::getVar('wpjobportal_nonce');
         if ( wp_verify_nonce( $nonce_value, 'wpjobportal_nonce') ) {
             if (isset($_POST['form_request']) && $_POST['form_request'] == 'wpjobportal')
                 return false;
             elseif (isset($_GET['action']) && $_GET['action'] == 'wpjobportaltask')
                 return false;
-            else
+            else{
+                if(!is_admin() && strpos($layout, 'admin_') === 0){
+                    return false;
+                }
                 return true;
+            }
         }
     }
 
@@ -54,6 +58,8 @@ class WPJOBPORTALHighesteducationController {
         if (! wp_verify_nonce( $nonce, 'wpjobportal_highest_education_nonce') ) {
              die( 'Security check Failed' );
         }
+        if(!wpjobportal::$_common->wpjp_isadmin())
+            return false;
         $data = WPJOBPORTALrequest::get('post');
         $result = WPJOBPORTALincluder::getJSModel('highesteducation')->storeHighestEducation($data);
         $url = esc_url_raw(admin_url('admin.php?page=wpjobportal_highesteducation&wpjobportallt=highesteducations'));
@@ -68,6 +74,8 @@ class WPJOBPORTALHighesteducationController {
         if (! wp_verify_nonce( $nonce, 'wpjobportal_highest_education_nonce') ) {
              die( 'Security check Failed' );
         }
+        if(!wpjobportal::$_common->wpjp_isadmin())
+            return false;
         $ids = WPJOBPORTALrequest::getVar('wpjobportal-cb');
         $result = WPJOBPORTALincluder::getJSModel('highesteducation')->deleteHighestEducations($ids);
         $msg = WPJOBPORTALMessages::getMessage($result, 'highesteducation');
@@ -82,6 +90,8 @@ class WPJOBPORTALHighesteducationController {
         if (! wp_verify_nonce( $nonce, 'wpjobportal_highest_education_nonce') ) {
              die( 'Security check Failed' );
         }
+        if(!wpjobportal::$_common->wpjp_isadmin())
+            return false;
         if(!wpjobportal::$_common->wpjp_isadmin()) return;
         $ids = WPJOBPORTALrequest::getVar('wpjobportal-cb');
         $pagenum = WPJOBPORTALrequest::getVar('pagenum');
@@ -100,6 +110,8 @@ class WPJOBPORTALHighesteducationController {
         if (! wp_verify_nonce( $nonce, 'wpjobportal_highest_education_nonce') ) {
              die( 'Security check Failed' );
         }
+        if(!wpjobportal::$_common->wpjp_isadmin())
+            return false;
         if(!wpjobportal::$_common->wpjp_isadmin()) return;
         $ids = WPJOBPORTALrequest::getVar('wpjobportal-cb');
         $pagenum = WPJOBPORTALrequest::getVar('pagenum');
@@ -119,6 +131,8 @@ class WPJOBPORTALHighesteducationController {
         if (! wp_verify_nonce( $nonce, 'wpjobportal_highest_education_nonce') ) {
              die( 'Security check Failed' );
         }
+        if(!wpjobportal::$_common->wpjp_isadmin())
+            return false;
         $post = WPJOBPORTALrequest::get('post');
         if($post['task'] == 'unpublish'){
             $this->unpublish();
